+++ old-ui/app/app.js
@@ -38,6 +38,7 @@ const AccountDropdowns = require('./components/account-dropdowns').AccountDropdo
 const DeleteRpc = require('./components/delete-rpc')
 const DeleteImportedAccount = require('./components/delete-imported-account')
 const ConfirmChangePassword = require('./components/confirm-change-password')
+const ethNetProps = require('eth-net-props')
 
 module.exports = connect(mapStateToProps)(App)
 
@@ -289,7 +290,7 @@ App.prototype.renderNetworkDropdown = function () {
         },
       },
       [h(providerType === 'poa' ? 'div.selected-network' : ''),
-        'POA Network',
+        ethNetProps.props.getNetworkDisplayName(99),
       ]
     ),
 
@@ -306,7 +307,7 @@ App.prototype.renderNetworkDropdown = function () {
         },
       },
       [h(providerType === 'sokol' ? 'div.selected-network' : ''),
-        'POA Sokol Test Network',
+        ethNetProps.props.getNetworkDisplayName(77),
       ]
     ),
 
@@ -323,7 +324,7 @@ App.prototype.renderNetworkDropdown = function () {
         },
       },
       [h(providerType === 'mainnet' ? 'div.selected-network' : ''),
-        'Main Ethereum Network',
+        ethNetProps.props.getNetworkDisplayName(1),
       ]
     ),
 
@@ -340,7 +341,7 @@ App.prototype.renderNetworkDropdown = function () {
         },
       },
       [h(providerType === 'ropsten' ? 'div.selected-network' : ''),
-        'Ropsten Test Network',
+        ethNetProps.props.getNetworkDisplayName(3),
       ]
     ),
 
@@ -357,7 +358,7 @@ App.prototype.renderNetworkDropdown = function () {
         },
       },
       [h(providerType === 'kovan' ? 'div.selected-network' : ''),
-        'Kovan Test Network',
+        ethNetProps.props.getNetworkDisplayName(42),
       ]
     ),
 
@@ -374,7 +375,7 @@ App.prototype.renderNetworkDropdown = function () {
         },
       },
       [h(providerType === 'rinkeby' ? 'div.selected-network' : ''),
-        'Rinkeby Test Network',
+        ethNetProps.props.getNetworkDisplayName(4),
       ]
     ),
 
@@ -406,7 +407,7 @@ App.prototype.renderNetworkDropdown = function () {
         style: {
           paddingLeft: '20px',
           fontSize: '16px',
-          color: '#8fdc97',
+          color: '#60db97',
         },
       },
       [
@@ -708,7 +709,7 @@ App.prototype.renderCustomOption = function (provider) {
           },
         },
         [h('div.selected-network'),
-          label,
+          h('.span.custom-rpc', label),
           h('.remove', {
             onClick: (event) => {
               event.preventDefault()
@@ -723,28 +724,8 @@ App.prototype.renderCustomOption = function (provider) {
 }
 
 App.prototype.getNetworkName = function () {
-  const { provider } = this.props
-  const providerName = provider.type
-
-  let name
-
-  if (providerName === 'mainnet') {
-    name = 'Main Ethereum Network'
-  } else if (providerName === 'sokol') {
-    name = 'POA Sokol Test Network'
-  } else if (providerName === 'ropsten') {
-    name = 'Ropsten Test Network'
-  } else if (providerName === 'kovan') {
-    name = 'Kovan Test Network'
-  } else if (providerName === 'rinkeby') {
-    name = 'Rinkeby Test Network'
-  } else if (providerName === 'poa') {
-    name = 'POA Network'
-  } else {
-    name = 'Unknown Private Network'
-  }
-
-  return name
+  const { network } = this.props
+  return ethNetProps.props.getNetworkDisplayName(network)
 }
 
 App.prototype.renderCommonRpc = function (rpcList, provider) {
@@ -767,7 +748,7 @@ App.prototype.renderCommonRpc = function (rpcList, provider) {
           },
         },
         [
-          rpc,
+          h('.span.custom-rpc', rpc),
           h('.remove', {
             onClick: (event) => {
               event.preventDefault()